It is an Rpg like game, with no graphics, in C++.
It is an Rpg like game, with no graphics. It is implemented in C++.
Its main purpose was to put to use basic inheritance logic, overloading
functions and other functionalities an object oriented programming
language
offers.
The game world is an 8x8 grid with monsters to fight, lakes forrests and
roads
to cross. The goal is to take your hero or heroes (1-3 allowed) from
point (0,0)
to (7,7). Heroes ,as monsters, have different types (warrior,mage etc)
attributes (health, speed, damage, stamina etc) and depending on its
class
an action being made may have different results (overloading functions).

Hint: If you are to play the game, select at least one warrior, that
category
is on god mode :P